requirements.txt
paho-mqtt
fastapi
uvicorn
psycopg2-binary
requests
python-dotenv
bleach
google-genai
